requirements.txt
imapclient
pyzmail
pandas
python-dotenv
google-generativeai
